chore: 🔧 clean up directory
