+++ iOS/issue-tracker/issue-tracker/Supporting Files/AppDelegate.swift
@@ -1,0 +1,18 @@
+//
+//  AppDelegate.swift
+//  issue-tracker
+//
+//  Created by 양준혁 on 2021/06/07.
+//
+
+import UIKit
+
+@main
+class AppDelegate: UIResponder, UIApplicationDelegate {
+    
+    func application(_ application: UIApplication, didFinishLaunchingWithOptions launchOptions: [UIApplication.LaunchOptionsKey: Any]?) -> Bool {
+        // Override point for customization after application launch.
+        return true
+    }
+}
+
