+++ iOS/issue-tracker/issue-tracker/Supporting Files/Assets.xcassets/AccentColor.colorset/Contents.json
