+++ iOS/issue-tracker/issue-tracker/Supporting Files/Assets.xcassets/AppIcon.appiconset/Contents.json
