+++ iOS/issue-tracker/issue-tracker/Supporting Files/Assets.xcassets/Contents.json
